ts: Move more code around
